Support for dimming inactive panes
Most of the changes are based on this Stack Overflow thread:
https://stackoverflow.com/a/33553372/734153

tmux already supports specifying different colors for inactive and
inactive panes so I used those and set them to the term colors.
Since the panes themselves have different color background, it doesn't
make sense to use borders so I set them to the background color to make
them "invisible".

vim was a lot trickier and required not setting the bg attributes in the
highlight group itself. I tried unsetting them, by clearing Normal and
then only setting the fg groups, but that didn't work. So I updated the
base16-colorscheme directly to get it to work.

A lot of the vim changes were fixes for the code relying on the fact
that utils#GetHighlightInfo(...) will always return the desired key and
not explicitly checking it.

Other changes:
* The base16_*_bg highlight groups were unnecessary and I deleted them
* I wasn't specifying 'MarkPalette solarized' anywhere.
  (No wonder vim-mark highlighting didn't work)