Split out code to tangle from one generating links
Moved all generated files to its subdir. This allows me to make that
directory be the generated branch to allow me to easily update it
Moved code from the main Makefile that creates links to the generated
files to a Makefile in the generated dir.
This allows me to just clone the generated branch on systems on which I
don't have access to emacs and use the Makefile to simply update links
Since I now have a Makefile which only updates links, it doesn't make
sense to do that in the main Makefile so the main Makefile simply calls
the generated Makefile